moved stencil.py
