+++ dictos/discrete/stencil.py
